+++ src/types/form-element.ts
@@ -204,6 +204,11 @@ export interface CodeOptions {
    * @type {string}
    */
   value: string;
+
+  /**
+   * Should escape value
+   */
+  isEscaping?: boolean;
 }
 
 /**
@@ -308,6 +313,11 @@ export interface TextareaOptions {
    * @type {string}
    */
   value: string;
+
+  /**
+   * Should escape value
+   */
+  isEscaping?: boolean;
 }
 
 /**